feat: Migrated Bookmarks Module to KMP
