+++ feature/bookmarks/src/commonMain/composeResources/drawable/feature_bookmarks_img_empty_bookmarks.xml
