+++ feature/bookmarks/src/commonMain/kotlin/com/google/samples/apps/nowinandroid/feature/bookmarks/BookmarksScreen.kt
@@ -1,0 +1,294 @@
+/*
+ * Copyright 2022 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.google.samples.apps.nowinandroid.feature.bookmarks
+
+import androidx.annotation.VisibleForTesting
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.gestures.Orientation
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.WindowInsets
+import androidx.compose.foundation.layout.fillMaxHeight
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.safeDrawing
+import androidx.compose.foundation.layout.systemBars
+import androidx.compose.foundation.layout.windowInsetsBottomHeight
+import androidx.compose.foundation.layout.windowInsetsPadding
+import androidx.compose.foundation.layout.wrapContentSize
+import androidx.compose.foundation.lazy.staggeredgrid.LazyVerticalStaggeredGrid
+import androidx.compose.foundation.lazy.staggeredgrid.StaggeredGridCells
+import androidx.compose.foundation.lazy.staggeredgrid.StaggeredGridItemSpan
+import androidx.compose.foundation.lazy.staggeredgrid.rememberLazyStaggeredGridState
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.getValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.ColorFilter
+import androidx.compose.ui.platform.testTag
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.unit.dp
+import androidx.lifecycle.Lifecycle
+import androidx.lifecycle.compose.LifecycleEventEffect
+import androidx.lifecycle.compose.collectAsStateWithLifecycle
+import bookmarks.generated.resources.Res
+import bookmarks.generated.resources.feature_bookmarks_empty_description
+import bookmarks.generated.resources.feature_bookmarks_empty_error
+import bookmarks.generated.resources.feature_bookmarks_img_empty_bookmarks
+import bookmarks.generated.resources.feature_bookmarks_loading
+import bookmarks.generated.resources.feature_bookmarks_removed
+import bookmarks.generated.resources.feature_bookmarks_undo
+import com.google.samples.apps.nowinandroid.core.designsystem.component.NiaLoadingWheel
+import com.google.samples.apps.nowinandroid.core.designsystem.component.scrollbar.DraggableScrollbar
+import com.google.samples.apps.nowinandroid.core.designsystem.component.scrollbar.rememberDraggableScroller
+import com.google.samples.apps.nowinandroid.core.designsystem.component.scrollbar.scrollbarState
+import com.google.samples.apps.nowinandroid.core.designsystem.theme.LocalTintTheme
+import com.google.samples.apps.nowinandroid.core.designsystem.theme.NiaTheme
+import com.google.samples.apps.nowinandroid.core.model.data.UserNewsResource
+import com.google.samples.apps.nowinandroid.core.ui.NewsFeedUiState
+import com.google.samples.apps.nowinandroid.core.ui.NewsFeedUiState.Loading
+import com.google.samples.apps.nowinandroid.core.ui.NewsFeedUiState.Success
+import com.google.samples.apps.nowinandroid.core.ui.TrackScreenViewEvent
+import com.google.samples.apps.nowinandroid.core.ui.TrackScrollJank
+import com.google.samples.apps.nowinandroid.core.ui.UserNewsResourcePreviewParameterProvider
+import com.google.samples.apps.nowinandroid.core.ui.newsFeed
+import org.jetbrains.compose.resources.painterResource
+import org.jetbrains.compose.resources.stringResource
+import org.jetbrains.compose.ui.tooling.preview.Preview
+import org.jetbrains.compose.ui.tooling.preview.PreviewParameter
+import org.koin.compose.viewmodel.koinViewModel
+
+@Composable
+internal fun BookmarksRoute(
+    onTopicClick: (String) -> Unit,
+    onShowSnackbar: suspend (String, String?) -> Boolean,
+    modifier: Modifier = Modifier,
+    viewModel: BookmarksViewModel = koinViewModel(),
+) {
+    val feedState by viewModel.feedUiState.collectAsStateWithLifecycle()
+    BookmarksScreen(
+        feedState = feedState,
+        onShowSnackbar = onShowSnackbar,
+        removeFromBookmarks = viewModel::removeFromSavedResources,
+        onNewsResourceViewed = { viewModel.setNewsResourceViewed(it, true) },
+        onTopicClick = onTopicClick,
+        modifier = modifier,
+        shouldDisplayUndoBookmark = viewModel.shouldDisplayUndoBookmark,
+        undoBookmarkRemoval = viewModel::undoBookmarkRemoval,
+        clearUndoState = viewModel::clearUndoState,
+    )
+}
+
+/**
+ * Displays the user's bookmarked articles. Includes support for loading and empty states.
+ */
+@VisibleForTesting(otherwise = VisibleForTesting.PRIVATE)
+@Composable
+internal fun BookmarksScreen(
+    feedState: NewsFeedUiState,
+    onShowSnackbar: suspend (String, String?) -> Boolean,
+    removeFromBookmarks: (String) -> Unit,
+    onNewsResourceViewed: (String) -> Unit,
+    onTopicClick: (String) -> Unit,
+    modifier: Modifier = Modifier,
+    shouldDisplayUndoBookmark: Boolean = false,
+    undoBookmarkRemoval: () -> Unit = {},
+    clearUndoState: () -> Unit = {},
+) {
+    val bookmarkRemovedMessage = stringResource(Res.string.feature_bookmarks_removed)
+    val undoText = stringResource(Res.string.feature_bookmarks_undo)
+
+    LaunchedEffect(shouldDisplayUndoBookmark) {
+        if (shouldDisplayUndoBookmark) {
+            val snackBarResult = onShowSnackbar(bookmarkRemovedMessage, undoText)
+            if (snackBarResult) {
+                undoBookmarkRemoval()
+            } else {
+                clearUndoState()
+            }
+        }
+    }
+
+    LifecycleEventEffect(Lifecycle.Event.ON_STOP) {
+        clearUndoState()
+    }
+
+    when (feedState) {
+        Loading -> LoadingState(modifier)
+        is Success -> if (feedState.feed.isNotEmpty()) {
+            BookmarksGrid(
+                feedState,
+                removeFromBookmarks,
+                onNewsResourceViewed,
+                onTopicClick,
+                modifier,
+            )
+        } else {
+            EmptyState(modifier)
+        }
+    }
+
+    TrackScreenViewEvent(screenName = "Saved")
+}
+
+@Composable
+private fun LoadingState(modifier: Modifier = Modifier) {
+    NiaLoadingWheel(
+        modifier = modifier
+            .fillMaxWidth()
+            .wrapContentSize()
+            .testTag("forYou:loading"),
+        contentDesc = stringResource(Res.string.feature_bookmarks_loading),
+    )
+}
+
+@Composable
+private fun BookmarksGrid(
+    feedState: NewsFeedUiState,
+    removeFromBookmarks: (String) -> Unit,
+    onNewsResourceViewed: (String) -> Unit,
+    onTopicClick: (String) -> Unit,
+    modifier: Modifier = Modifier,
+) {
+    val scrollableState = rememberLazyStaggeredGridState()
+    TrackScrollJank(scrollableState = scrollableState, stateName = "bookmarks:grid")
+    Box(
+        modifier = modifier
+            .fillMaxSize(),
+    ) {
+        LazyVerticalStaggeredGrid(
+            columns = StaggeredGridCells.Adaptive(300.dp),
+            contentPadding = PaddingValues(16.dp),
+            horizontalArrangement = Arrangement.spacedBy(16.dp),
+            verticalItemSpacing = 24.dp,
+            state = scrollableState,
+            modifier = Modifier
+                .fillMaxSize()
+                .testTag("bookmarks:feed"),
+        ) {
+            newsFeed(
+                feedState = feedState,
+                onNewsResourcesCheckedChanged = { id, _ -> removeFromBookmarks(id) },
+                onNewsResourceViewed = onNewsResourceViewed,
+                onTopicClick = onTopicClick,
+            )
+            item(span = StaggeredGridItemSpan.FullLine) {
+                Spacer(Modifier.windowInsetsBottomHeight(WindowInsets.safeDrawing))
+            }
+        }
+        val itemsAvailable = when (feedState) {
+            Loading -> 1
+            is Success -> feedState.feed.size
+        }
+        val scrollbarState = scrollableState.scrollbarState(
+            itemsAvailable = itemsAvailable,
+        )
+        scrollableState.DraggableScrollbar(
+            modifier = Modifier
+                .fillMaxHeight()
+                .windowInsetsPadding(WindowInsets.systemBars)
+                .padding(horizontal = 2.dp)
+                .align(Alignment.CenterEnd),
+            state = scrollbarState,
+            orientation = Orientation.Vertical,
+            onThumbMoved = scrollableState.rememberDraggableScroller(
+                itemsAvailable = itemsAvailable,
+            ),
+        )
+    }
+}
+
+@Composable
+private fun EmptyState(modifier: Modifier = Modifier) {
+    Column(
+        modifier = modifier
+            .padding(16.dp)
+            .fillMaxSize()
+            .testTag("bookmarks:empty"),
+        verticalArrangement = Arrangement.Center,
+        horizontalAlignment = Alignment.CenterHorizontally,
+    ) {
+        val iconTint = LocalTintTheme.current.iconTint
+        Image(
+            modifier = Modifier.fillMaxWidth(),
+            painter = painterResource(Res.drawable.feature_bookmarks_img_empty_bookmarks),
+            colorFilter = if (iconTint != Color.Unspecified) ColorFilter.tint(iconTint) else null,
+            contentDescription = null,
+        )
+
+        Spacer(modifier = Modifier.height(48.dp))
+
+        Text(
+            text = stringResource(Res.string.feature_bookmarks_empty_error),
+            modifier = Modifier.fillMaxWidth(),
+            textAlign = TextAlign.Center,
+            style = MaterialTheme.typography.titleMedium,
+            fontWeight = FontWeight.Bold,
+        )
+
+        Spacer(modifier = Modifier.height(8.dp))
+
+        Text(
+            text = stringResource(Res.string.feature_bookmarks_empty_description),
+            modifier = Modifier.fillMaxWidth(),
+            textAlign = TextAlign.Center,
+            style = MaterialTheme.typography.bodyMedium,
+        )
+    }
+}
+
+@Preview
+@Composable
+private fun LoadingStatePreview() {
+    NiaTheme {
+        LoadingState()
+    }
+}
+
+@Preview
+@Composable
+private fun BookmarksGridPreview(
+    @PreviewParameter(UserNewsResourcePreviewParameterProvider::class)
+    userNewsResources: List<UserNewsResource>,
+) {
+    NiaTheme {
+        BookmarksGrid(
+            feedState = Success(userNewsResources),
+            removeFromBookmarks = {},
+            onNewsResourceViewed = {},
+            onTopicClick = {},
+        )
+    }
+}
+
+@Preview
+@Composable
+private fun EmptyStatePreview() {
+    NiaTheme {
+        EmptyState()
+    }
+}
