+++ feature/bookmarks/src/commonTest/kotlin/com/google/samples/apps/nowinandroid/feature/bookmarks/BookmarksViewModelTest.kt
@@ -1,0 +1,134 @@
+/*
+ * Copyright 2022 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+@file:OptIn(ExperimentalCoroutinesApi::class)
+
+package com.google.samples.apps.nowinandroid.feature.bookmarks
+
+import com.google.samples.apps.nowinandroid.core.data.repository.CompositeUserNewsResourceRepository
+import com.google.samples.apps.nowinandroid.core.testing.data.newsResourcesTestData
+import com.google.samples.apps.nowinandroid.core.testing.repository.TestNewsRepository
+import com.google.samples.apps.nowinandroid.core.testing.repository.TestUserDataRepository
+import com.google.samples.apps.nowinandroid.core.ui.NewsFeedUiState.Loading
+import com.google.samples.apps.nowinandroid.core.ui.NewsFeedUiState.Success
+import kotlinx.coroutines.ExperimentalCoroutinesApi
+import kotlinx.coroutines.flow.collect
+import kotlinx.coroutines.launch
+import kotlinx.coroutines.test.UnconfinedTestDispatcher
+import kotlinx.coroutines.test.runTest
+import kotlin.test.BeforeTest
+import kotlin.test.Test
+import kotlin.test.assertEquals
+import kotlin.test.assertFalse
+import kotlin.test.assertIs
+import kotlin.test.assertTrue
+
+/**
+ * To learn more about how this test handles Flows created with stateIn, see
+ * https://developer.android.com/kotlin/flow/test#statein
+ */
+class BookmarksViewModelTest {
+    private val userDataRepository = TestUserDataRepository()
+    private val newsRepository = TestNewsRepository()
+    private val userNewsResourceRepository = CompositeUserNewsResourceRepository(
+        newsRepository = newsRepository,
+        userDataRepository = userDataRepository,
+    )
+    private lateinit var viewModel: BookmarksViewModel
+
+    @BeforeTest
+    fun setup() {
+        viewModel = BookmarksViewModel(
+            userDataRepository = userDataRepository,
+            userNewsResourceRepository = userNewsResourceRepository,
+        )
+    }
+
+    @Test
+    fun stateIsInitiallyLoading() = runTest {
+        assertEquals(Loading, viewModel.feedUiState.value)
+    }
+
+    @Test
+    fun oneBookmark_showsInFeed() = runTest {
+        backgroundScope.launch(UnconfinedTestDispatcher()) { viewModel.feedUiState.collect() }
+
+        newsRepository.sendNewsResources(newsResourcesTestData)
+        userDataRepository.setNewsResourceBookmarked(newsResourcesTestData[0].id, true)
+        val item = viewModel.feedUiState.value
+        assertIs<Success>(item)
+        assertEquals(item.feed.size, 1)
+    }
+
+    @Test
+    fun oneBookmark_whenRemoving_removesFromFeed() = runTest {
+        backgroundScope.launch(UnconfinedTestDispatcher()) { viewModel.feedUiState.collect() }
+        // Set the news resources to be used by this test
+        newsRepository.sendNewsResources(newsResourcesTestData)
+        // Start with the resource saved
+        userDataRepository.setNewsResourceBookmarked(newsResourcesTestData[0].id, true)
+        // Use viewModel to remove saved resource
+        viewModel.removeFromSavedResources(newsResourcesTestData[0].id)
+        // Verify list of saved resources is now empty
+        val item = viewModel.feedUiState.value
+        assertIs<Success>(item)
+        assertEquals(item.feed.size, 0)
+        assertTrue(viewModel.shouldDisplayUndoBookmark)
+    }
+
+    @Test
+    fun feedUiState_resourceIsViewed_setResourcesViewed() = runTest {
+        backgroundScope.launch(UnconfinedTestDispatcher()) { viewModel.feedUiState.collect() }
+
+        // Given
+        newsRepository.sendNewsResources(newsResourcesTestData)
+        userDataRepository.setNewsResourceBookmarked(newsResourcesTestData[0].id, true)
+        val itemBeforeViewed = viewModel.feedUiState.value
+        assertIs<Success>(itemBeforeViewed)
+        assertFalse(itemBeforeViewed.feed.first().hasBeenViewed)
+
+        // When
+        viewModel.setNewsResourceViewed(newsResourcesTestData[0].id, true)
+
+        // Then
+        val item = viewModel.feedUiState.value
+        assertIs<Success>(item)
+        assertTrue(item.feed.first().hasBeenViewed)
+    }
+
+    @Test
+    fun feedUiState_undoneBookmarkRemoval_bookmarkIsRestored() = runTest {
+        backgroundScope.launch(UnconfinedTestDispatcher()) { viewModel.feedUiState.collect() }
+
+        // Given
+        newsRepository.sendNewsResources(newsResourcesTestData)
+        userDataRepository.setNewsResourceBookmarked(newsResourcesTestData[0].id, true)
+        viewModel.removeFromSavedResources(newsResourcesTestData[0].id)
+        assertTrue(viewModel.shouldDisplayUndoBookmark)
+        val itemBeforeUndo = viewModel.feedUiState.value
+        assertIs<Success>(itemBeforeUndo)
+        assertEquals(0, itemBeforeUndo.feed.size)
+
+        // When
+        viewModel.undoBookmarkRemoval()
+
+        // Then
+        assertFalse(viewModel.shouldDisplayUndoBookmark)
+        val item = viewModel.feedUiState.value
+        assertIs<Success>(item)
+        assertEquals(1, item.feed.size)
+    }
+}
